Buffer size fixed

+++ src/aligner/aligner_containers.h
@@ -38,7 +38,7 @@ enum class AlignmentType {  // Strongly typed enum, C++11 feature.
 class AlignmentPosition {
  public:
   AlignmentPosition() : qstart(0), qend(0), tstart(0), tend(0) { }
-  AlignmentPosition(int32_t _qstart, int32_t _qend, int32_t _tstart, int32_t _tend) :
+  AlignmentPosition(int64_t _qstart, int64_t _qend, int64_t _tstart, int64_t _tend) :
                     qstart(_qstart), qend(_qend), tstart(_tstart), tend(_tend) { }
   AlignmentPosition(const AlignmentPosition& op) :
                     AlignmentPosition(op.qstart, op.qend, op.tstart, op.tend) { }
@@ -50,8 +50,8 @@ class AlignmentPosition {
     return *this;
   }
 
-  int32_t qstart, qend;              // Query and target alignment start and end positions. End position
-  int32_t tstart, tend;              // is inclusive (the position of the last base).
+  int64_t qstart, qend;              // Query and target alignment start and end positions. End position
+  int64_t tstart, tend;              // is inclusive (the position of the last base).
 };
 
 class AlignmentResult {
