+++ src/aligner/anchor_aligner.h
@@ -38,7 +38,7 @@ class AnchorAligner {
 
   ~AnchorAligner();
 
-  std::shared_ptr<AlignmentResult> CreateAlignmentResult(int qstart, int qend, int rstart, int rend, std::vector<is::CigarOp> rez);
+  std::shared_ptr<AlignmentResult> CreateAlignmentResult(int64_t qstart, int64_t qend, int64_t rstart, int64_t rend, std::vector<is::CigarOp> rez);
 
   void AdjustEnds(int left_offset_ref, int right_offset_ref, const char *query, const char *ref, int64_t *start_position_ref, int64_t *start_position_read, int number_of_bases, std::stack<is::CigarOp> *cigar_stack, std::deque<is::CigarOp> *cigar_queue, bool type);
